requirements.txt
asgiref==3.7.2
bleach==6.1.0
boto3==1.34.117
botocore==1.34.117
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==5.0.3
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf_api_logger==1.1.15
idna==3.7
jmespath==1.0.1
numpy==1.26.4
oauthlib==3.2.2
psycopg==3.1.12
psycopg-binary==3.1.12
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.1
urllib3==2.2.1
webencodings==0.5.1
watchtower==3.2.0
